requirements.txt
streamlit
pandas
numpy
scikit-learn
psycopg2-binary
openai
python-dotenv